Split trace end to end test by API
